+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - iterator.all
 - iterator.any
+- iterator.position
 - string.size
 - strings can now be used with the ordered comparison operators.